requirements.txt
aiohttp==3.6.2
aiopg==0.16.0
beautifulsoup4==4.8.1
peewee==2.10.2
peewee-async==0.5.12
python-dotenv==0.10.3

